Delete Recipes.xlsx
It's not recommended to upload excel files (with a file extension ".xlsx") into Git repository as Git treats it as a binary file (.exe file). And you can not  preview of the file as the Git is more tuned to read ".txt" file formats. The ".xlsx" format is a proprietary file format created by Microsoft for excel files. You should export the excel file into ".CSV" (Comma Separated Values) format and then upload it to Github.